Remove old interpreter
