+++ compiler/server/Server/Handlers.hs
@@ -38,7 +38,6 @@ import qualified Data.Set as S
 import Data.Text (Text)
 import GHC.Generics
 import qualified Language.Mimsa.Actions.Monad as Actions
-import Language.Mimsa.Interpreter.UseSwaps
 import Language.Mimsa.Printer
 import Language.Mimsa.Project.Helpers
 import Language.Mimsa.Project.Persistence
@@ -47,6 +46,7 @@ import Language.Mimsa.Project.Versions
 import Language.Mimsa.Store
 import Language.Mimsa.Tests.Test
 import Language.Mimsa.Tests.Types
+import Language.Mimsa.Typechecker.UseSwaps
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Identifiers
