+++ compiler/src/Language/Mimsa/Actions/Helpers/Swaps.hs
@@ -4,9 +4,8 @@ module Language.Mimsa.Actions.Helpers.Swaps
 where
 
 import Control.Monad.Except
-import Data.Functor
 import Language.Mimsa.Actions.Types
-import qualified Language.Mimsa.Interpreter.UseSwaps as Swaps
+import qualified Language.Mimsa.Typechecker.UseSwaps as Swaps
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Identifiers
@@ -19,4 +18,4 @@ useSwaps ::
 useSwaps swaps expr =
   case Swaps.useSwaps swaps expr of
     Right tyExpr -> pure tyExpr
-    Left e -> throwError (InterpreterErr (e $> mempty))
+    Left e -> throwError (TypeErr mempty e)
