+++ compiler/src/Language/Mimsa/Interpreter/App.hs
@@ -1,7 +1,7 @@
-module Language.Mimsa.Interpreter2.App (interpretApp) where
+module Language.Mimsa.Interpreter.App (interpretApp) where
 
-import Language.Mimsa.Interpreter2.Monad
-import Language.Mimsa.Interpreter2.Types
+import Language.Mimsa.Interpreter.Monad
+import Language.Mimsa.Interpreter.Types
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Interpreter.Stack
