+++ compiler/src/Language/Mimsa/Interpreter/Types.hs
@@ -1,130 +1,37 @@
 {-# LANGUAGE DerivingStrategies #-}
-{-# LANGUAGE GeneralisedNewtypeDeriving #-}
 
 module Language.Mimsa.Interpreter.Types
-  ( App (..),
-    InterpretState (..),
-    readScope,
-    nextVariable,
-    addToScope,
-    askForSwaps,
-    addOperator,
-    copySwap,
-    findOperator,
-    incrementApplyCount,
+  ( InterpreterM,
+    InterpretExpr,
+    InterpretFn,
+    InterpretReaderEnv (..),
+    InterpretPattern,
   )
 where
 
-import Control.Applicative
-import Control.Monad.Except
-import Control.Monad.State.Lazy
+import Control.Monad.Reader
 import Data.Map (Map)
-import qualified Data.Map as M
-import Data.Maybe (listToMaybe)
 import Language.Mimsa.Types.AST
-import Language.Mimsa.Types.Error
-import Language.Mimsa.Types.Identifiers
-import Language.Mimsa.Types.Scope
-import Language.Mimsa.Types.Swaps
+import Language.Mimsa.Types.Error.InterpreterError
+import Language.Mimsa.Types.Interpreter.Stack
+import Language.Mimsa.Types.Store.ExprHash
 
-newtype App ann a = App
-  { getApp ::
-      ExceptT
-        (InterpreterError ann)
-        (State (InterpretState ann))
-        a
+type InterpreterM var ann a =
+  ReaderT
+    (InterpretReaderEnv var ann)
+    (Either (InterpreterError var ann))
+    a
+
+data InterpretReaderEnv var ann = InterpretReaderEnv
+  { ireStack :: StackFrame var ann,
+    ireGlobals :: Map ExprHash (InterpretExpr var ann)
   }
-  deriving newtype
-    ( Functor,
-      Applicative,
-      Monad,
-      Alternative,
-      MonadState (InterpretState ann),
-      MonadError (InterpreterError ann)
-    )
 
-data InterpretState ann = InterpretState
-  { isVarNum :: Int,
-    isScope :: Scope ann,
-    isInfix :: Map InfixOp (Expr Variable ann),
-    -- number of applications we have done for timeout
-    isApplyCount :: Int,
-    isSwaps :: Swaps
-  }
-  deriving stock (Eq, Ord, Show)
+type InterpretExpr var ann = Expr (var, Maybe ExprHash) (ExprData var ann)
 
--- infix operators
+type InterpretPattern var ann =
+  Pattern (var, Maybe ExprHash) (ExprData var ann)
 
-addOperator :: InfixOp -> Expr Variable ann -> App ann ()
-addOperator infixOp expr = do
-  modify (\is -> is {isInfix = isInfix is <> M.singleton infixOp expr})
-
-findOperator :: InfixOp -> App ann (Maybe (Expr Variable ann))
-findOperator infixOp = do
-  ops <- gets isInfix
-  pure (M.lookup infixOp ops)
-
--- variable numbers
-
-nextInt :: App ann Int
-nextInt = do
-  int' <- gets isVarNum
-  modify (\is -> is {isVarNum = 1 + isVarNum is})
-  pure int'
-
-nextVariable :: App ann Variable
-nextVariable = NumberedVar <$> nextInt
-
--- scope
-
-readScope :: App ann (Scope ann)
-readScope = gets isScope
-
-addToScope :: (Eq ann, Monoid ann) => Scope ann -> App ann ()
-addToScope scope' =
-  case foundALoop scope' of
-    Nothing ->
-      modify
-        ( \is ->
-            is
-              { isVarNum = 1 + isVarNum is,
-                isScope = scope' <> isScope is
-              }
-        )
-    Just k -> throwError $ SelfReferencingBinding k
-  where
-    foundALoop (Scope newScope) =
-      fmap fst . listToMaybe . M.toList . M.filterWithKey (\k a -> MyVar mempty k == a) $ newScope
-
--- infinity protection
-
--- number of function applications before we fail
-maxCount :: Int
-maxCount = 500000
-
-incrementApplyCount :: App ann ()
-incrementApplyCount = do
-  appCount <- gets isApplyCount
-  if appCount < maxCount
-    then modify (\is -> is {isApplyCount = appCount + 1})
-    else throwError MaximumCallSizeReached
-
--- track original names of numbered vars for pretty printing later
-
-askForSwaps :: App ann Swaps
-askForSwaps = gets isSwaps
-
-addSwap :: Variable -> Name -> App ann ()
-addSwap v n = modify (\is -> is {isSwaps = isSwaps is <> M.singleton v n})
-
-findSwap :: Variable -> App ann (Maybe Name)
-findSwap v =
-  M.lookup v <$> askForSwaps
-
-copySwap :: Variable -> Variable -> App ann ()
-copySwap old new = do
-  name <- findSwap old
-  case name of
-    Nothing -> pure ()
-    Just foundSwap -> do
-      addSwap new foundSwap
+type InterpretFn var ann =
+  InterpretExpr var ann ->
+  InterpreterM var ann (InterpretExpr var ann)
