+++ compiler/test/Test/Backend/Typescript.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE TypeApplications #-}
 
 module Test.Backend.Typescript
   ( spec,
@@ -21,8 +22,8 @@ import Language.Mimsa.Backend.Typescript.Monad
 import Language.Mimsa.Backend.Typescript.Patterns
 import Language.Mimsa.Backend.Typescript.Printer
 import Language.Mimsa.Backend.Typescript.Types
-import Language.Mimsa.Interpreter.UseSwaps
 import Language.Mimsa.Printer
+import Language.Mimsa.Typechecker.UseSwaps
 import Language.Mimsa.Types.AST
 import Language.Mimsa.Types.Error
 import Language.Mimsa.Types.Identifiers
@@ -49,7 +50,7 @@ testFromInputText input =
     Right resolved -> do
       exprName <-
         first
-          (prettyPrint . InterpreterErr)
+          (prettyPrint . TypeErr @Annotation input)
           (useSwaps (reSwaps resolved) (reTypedExpression resolved))
       let readerState = TSReaderState mempty
       first prettyPrint (printModule . fst <$> fromExpr readerState exprName)
